+++ CHANGELOG.md
@@ -14,7 +14,8 @@ IMPROVEMENTS:
 
 BUG FIXES:
 
-- hashgraph: stopping-condition in updateAncestorsFirstDescendants.
+- hashgraph: Change stopping-condition in updateAncestorsFirstDescendants.
+- node: Change eviction-condition in checkSuspend.
 
 ## v0.5.12 (November 27, 2019)